Renamed Admin area to Data area and added controllers and Views for MenuItems and others. Updated Store and MenuItems model with correct data types and two new fields (in Store) to have the phone number and location

+++ FoodOrder.DataAccess/Model/MenuItem.cs
@@ -10,7 +10,7 @@ namespace FoodOrder.DataAccess.Model
 
         public virtual string Name { get; set; }
 
-        public virtual int Price { get; set; }
+        public virtual float Price { get; set; }
 
     }
